#init:
"Added new 'Todo' project and weather forecast feature"

Added a new project named "Todo" to the solution file `BaltaAspNet.sln` with configurations for Debug and Release modes. Two new configuration files `appsettings.Development.json` and `appsettings.json` were added for logging levels and allowed hosts. The `Program.cs` file was updated to setup a new web application. A new project file `Todo.csproj` was created with .NET 8.0 as the target framework and a package reference to Swashbuckle.AspNetCore version 6.4.0. A new HTTP file `Todo.http` was created for a GET request to the weather forecast endpoint. A new controller `WeatherForecastController.cs` and a new model `WeatherForecast.cs` were added to the Todo project for a weather forecast feature. The `launchSettings.json` file was updated for running the application in development mode with the launch URL set to "swagger".