#DB:
Add initial database setup and migration files

Added `AppDbContext.cs` which defines the `AppDbContext` class for interacting with the database and includes a `DbSet` for `TodoModel` objects. Also added `20240307172146_CreateDatabase.cs` and `20240307172146_CreateDatabase.Designer.cs` files which define the `CreateDatabase` migration for creating the initial database schema. The `AppDbContextModelSnapshot.cs` file was added to provide a snapshot of the current state of the model for determining changes in future migrations.