#Controller:
Created TodoController class and added basic functionality

Added the `Microsoft.AspNetCore.Mvc` and `Todo.Models` namespaces to the `TodoController.cs` file. Created a new `TodoController` class that inherits from `ControllerBase` and is decorated with the `ApiController` attribute. The `Route` attribute has been added to the `TodoController` class with the value "[controller]". A constructor for the `TodoController` class has been added, which takes an `ILogger<TodoController>` parameter and stores it in a private readonly field `_logger`. A `Get` method has been added to the `TodoController` class, decorated with the `HttpGet` attribute and a name of "GetTodo". This method logs an informational message and returns an empty list.